+++ Cargo.toml
@@ -26,20 +26,23 @@ clap = "2.33.3"
 unicode-width = "0.1.8"
 resize = "0.5.5"
 wavy = "0.9.0"
-fon = "0.5.0"
-pasts = "0.7.4"
 shellwords = "1.1.0"
 shellexpand = "2.1"
-futures = {version= "0.3.12", features = ["executor"]}
 toml = "0.5.8"
 dirs-next = "2.0.0"
 
-[target.'cfg(target_os = "linux")'.dependencies]
-v4l = { version = "0.12.0", optional = true }
 
 [dev-dependencies]
 rand = "0.8.1"
 
 [features]
 stream-video = ["v4l"]
+stream-audio = ["fon", "pasts", "futures"]
 ui-test = []
+
+[target.'cfg(target_os = "linux")'.dependencies]
+v4l = { version = "0.12.0", optional = true }
+fon = {version = "0.5.0", optional = true}
+pasts = {version = "0.7.4", optional = true}
+futures = {version= "0.3.12", features = ["executor"], optional=true}
+
