+++ src/state.rs
@@ -1,10 +1,12 @@
 use message_io::network::Endpoint;
 use chrono::{DateTime, Local};
-use fon::{stereo::Stereo32, Audio, Sink};
-use wavy::{Speakers, SpeakersSink};
 
 use std::collections::HashMap;
-use std::sync::mpsc;
+
+#[cfg(feature = "stream-audio")]
+mod audio;
+#[cfg(feature = "stream-audio")]
+use audio::AudioStream;
 
 #[derive(PartialEq)]
 pub enum SystemMessageType {
@@ -51,21 +53,6 @@ impl Window {
     }
 }
 
-pub struct AudioStream {
-    tx: mpsc::Sender<Vec<u8>>,
-    tx_stop: mpsc::Sender<()>,
-}
-
-impl AudioStream {
-    fn update(&self, audio: Vec<u8>) {
-        self.tx.send(audio).unwrap();
-    }
-
-    fn stop(&self) {
-        self.tx_stop.send(()).unwrap();
-    }
-}
-
 #[derive(Default)]
 pub struct State {
     messages: Vec<ChatMessage>,
@@ -78,6 +65,7 @@ pub struct State {
     pub stop_stream: bool,
     pub stop_audio: bool,
     pub windows: HashMap<Endpoint, Window>,
+    #[cfg(feature = "stream-audio")]
     pub audio: Option<AudioStream>,
 }
 
@@ -289,46 +277,5 @@ impl State {
         window.width = width;
         window.height = height;
         window.data = data;
-    }
-
-    pub fn pulse_audio(&mut self, audio: Vec<u8>) {
-        fn start_audio() -> AudioStream {
-            const SAMPLE_RATE: f32 = 48_000.;
-            let (tx, rx) = mpsc::channel();
-            let (tx_stop, rx_stop) = mpsc::channel();
-            std::thread::spawn(move || {
-                let mut speakers = Speakers::default();
-                let mut buffer: Audio<Stereo32> = Audio::with_silence(SAMPLE_RATE, 0);
-                futures::executor::block_on(async move {
-                    loop {
-                        let mut speakers: SpeakersSink<'_, Stereo32> = speakers.play().await;
-                        speakers.stream(buffer.drain());
-                        let data: Vec<u8> = rx.try_iter().flatten().collect();
-                        let data: Vec<f32> = data
-                            .chunks_exact(4)
-                            .map(|array| std::convert::TryFrom::try_from(array).unwrap())
-                            .map(f32::from_le_bytes)
-                            .collect();
-                        let mut audio: Audio<Stereo32> = Audio::with_f32_buffer(SAMPLE_RATE, data);
-                        buffer.extend(audio.drain());
-                        if rx_stop.try_recv().is_ok() {
-                            break
-                        }
-                    }
-                });
-            });
-            AudioStream { tx, tx_stop }
-        }
-        if self.audio.is_none() {
-            self.audio = Some(start_audio());
-        }
-        // safe unwrap
-        self.audio.as_ref().unwrap().update(audio);
-    }
-
-    pub fn stop_audio(&mut self) {
-        if let Some(audio) = self.audio.take() {
-            audio.stop();
-        }
     }
 }
